create React Apps sections
