+++ client/src/components/frame.jsx
@@ -1,12 +1,12 @@
 import React from 'react';
 import Header from '../components/Header.jsx';
-import About from '../components/About.jsx';
+import Portfolio from './portfolio/Portfolio.jsx';
 
 export default class Frame extends React.Component {
   render() {
     return <React.Fragment>
       <Header/>
-      <About/>
+      <Portfolio/>
     </React.Fragment>
   }
 }
