+++ client/src/components/portfolio/Portfolio.jsx
@@ -1,0 +1,13 @@
+import React from 'react';
+import About from './About.jsx';
+
+const Portfolio = () => {
+
+  return (
+    <React.Fragment>
+      <About/>
+    </React.Fragment>
+  )
+}
+
+export default Portfolio;
