+++ client/src/components/portfolio/react/YoutubeApp.jsx
@@ -1,0 +1,20 @@
+import React from 'react';
+
+const YoutubeApp = () => {
+  return (
+    <React.Fragment>
+      <h2><span><em>React.js (click images)</em></span></h2>
+      <section className="portfolio" >
+        <section className="thirds">
+          <a href="recastlyRedux.html">
+            <img src="img/reactApp.jpg" atl="" id="registry-img"/>
+            <p id="para-target">YouTube API React Redux App</p>
+          </a>
+
+        </section>
+      </section>
+    </React.Fragment>
+  )
+}
+
+export default YoutubeApp
