+++ client/src/index.jsx
@@ -1,5 +1,5 @@
 import React from 'react';
 import ReactDOM from 'react-dom'
-import Frame from './components/frame.jsx'
+import Frame from './components/Frame.jsx'
 
 ReactDOM.render(<Frame/>, document.getElementById('root'));
